wtimer initial commit
wtimer is a high performance timers implementation using a hierarchical
 timer wheel.

It is optimised for high number of timers (100k+) with relatively
lower precision requirements  (>20ms). Internal heap allocations
can be completely avoided if the timer handler structures are
provided by the caller and if logging is set to a lower level.